fix: split GdsPinDetector to respect 500-line architecture limit
Move edge/coordinate/interval helpers into GdsPinDetector.EdgeHelpers.cs
so the main detector file drops below the hard file-size limit enforced by
UnitTests.Architecture.FileSizeLimitTests.
